Extract WaitForVirtualAgentToken
