+++ src/applications/virtual-agent/containers/WaitForFeatureToggles.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import LoadingIndicator from '@department-of-veterans-affairs/component-library/LoadingIndicator';
 import { connect } from 'react-redux';
-import WebChat from './WebChat';
+import WaitForVirtualAgentToken from './WaitForVirtualAgentToken';
 
 function WaitForFeatureToggles({ featureTogglesLoading }) {
   if (featureTogglesLoading) {
     return <LoadingIndicator message={'Loading Virtual Agent'} />;
   }
-  return <WebChat />;
+  return <WaitForVirtualAgentToken />;
 }
 
 const mapStateToProps = state => {
